Move PreviewDatavis into its own file
