+++ src/blocks/datavis-block/EditDatavisBlock.js
@@ -18,7 +18,6 @@ import { __ } from '@wordpress/i18n';
 import ControlledJsonEditor from '../../components/ControlledJsonEditor';
 import DatasetEditor from '../../components/DatasetEditor';
 import VegaChart from '../../components/VegaChart';
-import { setupDatavisBlocks } from '../../index';
 
 import defaultSpecification from './specification.json';
 
@@ -165,11 +164,6 @@ const EditDatavisBlock = ( { attributes, setAttributes, isSelected } ) => {
 	const blockProps = useBlockProps();
 	const json = attributes.json || defaultSpecification;
 
-	const blockId = blockProps.id;
-	setAttributes( { blockId } );
-
-	PreviewDatavis( blockId );
-
 	return (
 		<div { ...blockProps }>
 			<VegaChart spec={ json } />
@@ -205,45 +199,6 @@ const EditDatavisBlock = ( { attributes, setAttributes, isSelected } ) => {
 			) : null }
 		</div>
 	);
-};
-
-/**
- * Hacky way to get the Datavis model to render after updates have been made to the block.
- *
- * This is needed due to the lack of an action when the Server Side render is completed. So we create a interval
- * to check if the block element has been updated, if not it run the Datavis setup.
- *
- * @param {string} blockId Block id.
- * @class
- */
-const PreviewDatavis = ( blockId ) => {
-	let dataVisLoaded = false,
-		killCounter = 0;
-	let dataVisLoadedInterval = setInterval( function() {
-		const element = document.getElementById( blockId+'-datavis' );
-		if ( ! element ) {
-			return;
-		}
-		// To prevent multiple reloads, only run preview if the block element has loaded and the Datavis has not been loaded.
-		let loadedElement = element.dataset.loaded ?? false,
-			loaded = ( loadedElement === 'true' );
-
-		if ( element && ! loaded ) {
-			// Element is present but the dataviz configuration has not yet rendered from the server.
-			dataVisLoaded = true;
-			element.setAttribute( 'data-loaded', true );
-		}
-		// Element exists and the dataviz configuration has rendered from the server so kick off the processing of the dataviz configuration.
-		if ( dataVisLoaded && loaded ) {
-			setupDatavisBlocks();
-			clearInterval( dataVisLoadedInterval );
-		}
-		// If after a period of time kill the interval to not fall into an infinite loop.
-		if ( killCounter > 10 ) {
-			clearInterval( dataVisLoadedInterval );
-		}
-		killCounter += 1;
-	}, 1000 );
 };
 
 export default EditDatavisBlock;
